Reorganize API code to contain the proliferating worker glue files
