+++ src/setupTests.js
@@ -13,5 +13,5 @@ globalThis.TextDecoder = TextDecoder;
 // Make sure the mock version of the web worker we use for the local API
 // implementation loads under Jest, where web workers are not actually
 // available.
-jest.mock("./LocalAPIWorkerFactory")
+jest.mock("./api/local/WorkerFactory")
